cdk/requirements.txt
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_ecs
aws-cdk.aws_ecr
aws-cdk.aws_iam
aws_cdk.aws_logs
aws-cdk.aws_elasticloadbalancingv2
aws-cdk.aws_route53
aws-cdk.aws_route53_targets